CB-28 Use cluster definition multi-delete
The "clusterdefinition delete" command is updated to allow for multiple
definition names to be provided. The CLI then uses the new multi-delete
endpoint for cluster definitions to perform the deletion (even if only
one name is provided). As with the prior version of the delete command,
it's an error to not supply any names.