CB-29 Use recipe multi-delete
The "recipe delete" command is updated to allow for multiple recipe
names to be provided. The CLI then uses the new multi-delete endpoint
for recipes to perform the deletion (even if only one name is provided).
As with the prior version of the delete command, it's an error to not
supply any names.