CB-1845 Complete redbeams dbserver commands
The following new commands are available for redbeams dbserver:

* create, for allocating new service-managed database servers
* register, for registering existing user-managed database servers

In addition:

The formerly separate describe / describe-by-name and get-status /
get-status-by-name redbeams operations are merged into single commands.
For either of the new commands, a valid call is one that passes either
only a CRN, or a pair of environment CRN and name. The command
implementations call the appropriate redbeams API endpoint depending
on the provided options.

New optional CRN flags are introduced, as well as a new function for
checking on valid "resource addressing" option combinations, in order
to enable this change.